feat: 리뷰 컨트롤러 및 테스트 추가

+++ src/main/java/com/triple/triplehomework/controller/ReviewController.java
@@ -8,6 +8,7 @@ import com.triple.triplehomework.service.ReviewService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.StringUtils;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -31,21 +32,35 @@ public class ReviewController {
         }
 
         final ReviewActionCode action = reviewRequestDto.getAction();
+        final boolean hasReviewId = StringUtils.hasText(reviewRequestDto.getReviewId());
+
         switch (action){
             case ADD:
-                final ReviewResponseDto registerBody = reviewService.register(reviewRequestDto);
-                return responseDto.ok(registerBody, "리뷰 작성 성공", HttpStatus.OK);
-            case MOD:
-                final ReviewResponseDto modifyBody = reviewService.modify(reviewRequestDto);
-                return responseDto.ok(modifyBody, "리뷰 수정 성공", HttpStatus.OK);
-            case DELETE:
-                final boolean removeBody = reviewService.remove(reviewRequestDto);
 
-                if(reviewRequestDto.getRemovePhotoYn().equalsIgnoreCase("Y")){
-                    return responseDto.ok(removeBody, "첨부 파일 삭제 성공", HttpStatus.OK);
+                if(StringUtils.hasText(reviewRequestDto.getContent())){
+                    final ReviewResponseDto registerBody = reviewService.register(reviewRequestDto);
+                    return responseDto.ok(registerBody, "리뷰 작성 성공", HttpStatus.OK);
+                }
+                return responseDto.badRequest("리뷰 내용은 비어있을 수 없습니다.");
+
+            case MOD:
+
+                if(hasReviewId){
+                    final ReviewResponseDto modifyBody = reviewService.modify(reviewRequestDto);
+                    return responseDto.ok(modifyBody, "리뷰 수정 성공", HttpStatus.OK);
+                }
+                return responseDto.badRequest("리뷰 ID는 필수값입니다.");
+
+            case DELETE:
+
+                if(hasReviewId){
+                    final boolean removeBody = reviewService.remove(reviewRequestDto);
+                    if(reviewRequestDto.getRemovePhotoYn().equalsIgnoreCase("Y")){
+                        return responseDto.ok(removeBody, "첨부 파일 삭제 성공", HttpStatus.OK);
+                    }
+                    return responseDto.ok(removeBody, "리뷰 삭제 성공", HttpStatus.OK);
                 }
 
-                return responseDto.ok(removeBody, "리뷰 삭제 성공", HttpStatus.OK);
             default:
                 return responseDto.badRequest("리뷰 요청 이벤트 에러");
         }
